Adds a Routing method to calculate cost.
This method currently calculates cost in a way that is usually right but sometimes
wrong -- namely, when a leg must be flown more than once, this cost function doesn't
account for it.  Nevertheless, that is a fiddly detail that can be added later, and
the existence of a cost function allows me to get started on the actual routing algorithm.